reforge dot multiplication

+++ scilab/modules/ast/includes/operations/types_subtraction.hxx
@@ -103,18 +103,18 @@ template<> inline types::InternalType* sub_M_M<types::Sparse, types::Double, typ
 
 //add matrix - matrix ( double, int, bool )
 //same type
-template<typename T, typename O> inline static void sub(T* l, long long size, T* r, O* o)
+template<typename T, typename O> inline static void sub(T* l, size_t size, T* r, O* o)
 {
-    for (int i = 0; i < size ; ++i)
+    for (size_t i = 0; i < size ; ++i)
     {
         o[i] = (O)l[i] - (O)r[i];
     }
 }
 
 //x - x
-template<typename T, typename U, typename O> inline static void sub(T* l, long long size, U* r, O* o)
+template<typename T, typename U, typename O> inline static void sub(T* l, size_t size, U* r, O* o)
 {
-    for (int i = 0; i < size ; ++i)
+    for (size_t i = 0; i < size ; ++i)
     {
         o[i] = (O)l[i] - (O)r[i];
     }
@@ -122,9 +122,9 @@ template<typename T, typename U, typename O> inline static void sub(T* l, long l
 
 
 //x - xC
-template<typename T, typename U, typename O> inline static void sub(T* l, long long size, U* r, U* rc, O* o, O* oc)
+template<typename T, typename U, typename O> inline static void sub(T* l, size_t size, U* r, U* rc, O* o, O* oc)
 {
-    for (int i = 0; i < size ; ++i)
+    for (size_t i = 0; i < size ; ++i)
     {
         o[i] = (O)l[i] - (O)r[i];
         oc[i] = (O) - rc[i];
@@ -132,9 +132,9 @@ template<typename T, typename U, typename O> inline static void sub(T* l, long l
 }
 
 //xC - x
-template<typename T, typename U, typename O> inline static void sub(T* l, T* lc, long long size, U* r, O* o, O* oc)
+template<typename T, typename U, typename O> inline static void sub(T* l, T* lc, size_t size, U* r, O* o, O* oc)
 {
-    for (int i = 0; i < size ; ++i)
+    for (size_t i = 0; i < size ; ++i)
     {
         o[i] = (O)l[i] - (O)r[i];
         oc[i] = (O)lc[i];
@@ -142,9 +142,9 @@ template<typename T, typename U, typename O> inline static void sub(T* l, T* lc,
 }
 
 // xC - xC
-template<typename T, typename O> inline static void sub(T* l, T* lc, long long size, T* r, T* rc, O* o, O* oc)
+template<typename T, typename O> inline static void sub(T* l, T* lc, size_t size, T* r, T* rc, O* o, O* oc)
 {
-    for (int i = 0; i < size ; ++i)
+    for (size_t i = 0; i < size ; ++i)
     {
         o[i] = (O)l[i] - (O)r[i];
         oc[i] = (O)lc[i] - (O)rc[i];
@@ -153,18 +153,18 @@ template<typename T, typename O> inline static void sub(T* l, T* lc, long long s
 
 //add matrix - scalar ( double, int, bool )
 //x - x1
-template<typename T, typename U, typename O> inline static void sub(T* l, long long size, U r, O* o)
+template<typename T, typename U, typename O> inline static void sub(T* l, size_t size, U r, O* o)
 {
-    for (int i = 0; i < size ; ++i)
+    for (size_t i = 0; i < size ; ++i)
     {
         o[i] = (O)l[i] - (O)r;
     }
 }
 
 //xC - x1
-template<typename T, typename U, typename O> inline static void sub(T* l, T* lc, long long size, U r, O* o, O* oc)
+template<typename T, typename U, typename O> inline static void sub(T* l, T* lc, size_t size, U r, O* o, O* oc)
 {
-    for (int i = 0; i < size ; ++i)
+    for (size_t i = 0; i < size ; ++i)
     {
         o[i] = (O)l[i] - (O)r;
         oc[i] = (O)lc[i];
@@ -172,9 +172,9 @@ template<typename T, typename U, typename O> inline static void sub(T* l, T* lc,
 }
 
 //x - x1C
-template<typename T, typename U, typename O> inline static void sub(T* l, long long size, U r, U rc, O* o, O* oc)
+template<typename T, typename U, typename O> inline static void sub(T* l, size_t size, U r, U rc, O* o, O* oc)
 {
-    for (int i = 0; i < size ; ++i)
+    for (size_t i = 0; i < size ; ++i)
     {
         o[i] = (O)l[i] - (O) r;
         oc[i] = (O) - rc;
@@ -182,9 +182,9 @@ template<typename T, typename U, typename O> inline static void sub(T* l, long l
 }
 
 //xC - x1C
-template<typename T, typename O> inline static void sub(T* l, T* lc, long long size, T r, T rc, O* o, O* oc)
+template<typename T, typename O> inline static void sub(T* l, T* lc, size_t size, T r, T rc, O* o, O* oc)
 {
-    for (int i = 0; i < size ; ++i)
+    for (size_t i = 0; i < size ; ++i)
     {
         o[i] = (O)l[i] - (O)r;
         oc[i] = (O)lc[i] - (O)rc;
@@ -193,18 +193,18 @@ template<typename T, typename O> inline static void sub(T* l, T* lc, long long s
 
 //add scalar - matrix ( double, int, bool )
 //x1 - x
-template<typename T, typename U, typename O> inline static void sub(T l, long long size, U* r, O* o)
+template<typename T, typename U, typename O> inline static void sub(T l, size_t size, U* r, O* o)
 {
-    for (int i = 0; i < size ; ++i)
+    for (size_t i = 0; i < size ; ++i)
     {
         o[i] = (O)l - (O)r[i];
     }
 }
 
 //x1 - xC
-template<typename T, typename U, typename O> inline static void sub(T l, long long size, U* r, U* rc, O* o, O* oc)
+template<typename T, typename U, typename O> inline static void sub(T l, size_t size, U* r, U* rc, O* o, O* oc)
 {
-    for (int i = 0; i < size ; ++i)
+    for (size_t i = 0; i < size ; ++i)
     {
         o[i] = (O)l - (O)r[i];
         oc[i] = (O) - rc[i];
@@ -212,9 +212,9 @@ template<typename T, typename U, typename O> inline static void sub(T l, long lo
 }
 
 //x1C - x
-template<typename T, typename U, typename O> inline static void sub(T l, T lc, long long size, U* r, O* o, O* oc)
+template<typename T, typename U, typename O> inline static void sub(T l, T lc, size_t size, U* r, O* o, O* oc)
 {
-    for (int i = 0; i < size ; ++i)
+    for (size_t i = 0; i < size ; ++i)
     {
         o[i] = (O)l - (O)r[i];
         oc[i] = (O)lc;
@@ -222,9 +222,9 @@ template<typename T, typename U, typename O> inline static void sub(T l, T lc, l
 }
 
 //x1C - xC
-template<typename T, typename O> inline static void sub(T l, T lc, long long size, T* r, T* rc, O* o, O* oc)
+template<typename T, typename O> inline static void sub(T l, T lc, size_t size, T* r, T* rc, O* o, O* oc)
 {
-    for (int i = 0; i < size ; ++i)
+    for (size_t i = 0; i < size ; ++i)
     {
         o[i] = (O)l - (O)r[i];
         oc[i] = (O)lc - rc[i];
